test(ts-sdk): test Batch and FungibleAssetOrder
