+++ ts-sdk/test/evm.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect } from "vitest"
 import { ucs03abi, instructionAbi } from "../src/evm/abi/index.js"
+import { Batch, FungibleAssetOrder } from "../src/evm/ucs03/index.js"
 
 describe("abi", () => {
   it("should export ucs03abi", () => {
@@ -20,5 +21,29 @@ describe("abi", () => {
 
     const argNames = instructionAbi.map(item => item.name)
     expect(argNames).toContain("operand")
+  })
+})
+
+describe("ucs03", () => {
+  it("should encode FungibleAssetOrder and Batch correctly", () => {
+    const fungibleAssetOrder = FungibleAssetOrder([
+      "0xE6831e169d77a861A0E71326AFA6d80bCC8Bc6aA",
+      "0xE6831e169d77a861A0E71326AFA6d80bCC8Bc6aA",
+      "0x74d5b8eacfeb0dadaaf66403f40e304b3ef968b3",
+      4n,
+      "muno",
+      "muno",
+      18,
+      0n,
+      "0x74d5b8eacfeb0dadaaf66403f40e304b3ef968b3",
+      4n
+    ])
+    const batch = Batch([fungibleAssetOrder, fungibleAssetOrder])
+    expect(batch).toStrictEqual({
+      opcode: 2,
+      version: 0,
+      operand:
+        "0x000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000038000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000002c00000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000001c000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000024000000000000000000000000000000000000000000000000000000000000000120000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000014E6831e169d77a861A0E71326AFA6d80bCC8Bc6aA0000000000000000000000000000000000000000000000000000000000000000000000000000000000000014E6831e169d77a861A0E71326AFA6d80bCC8Bc6aA000000000000000000000000000000000000000000000000000000000000000000000000000000000000001474d5b8eacfeb0dadaaf66403f40e304b3ef968b300000000000000000000000000000000000000000000000000000000000000000000000000000000000000046d756e6f0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000046d756e6f00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001474d5b8eacfeb0dadaaf66403f40e304b3ef968b300000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000002c00000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000001c000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000024000000000000000000000000000000000000000000000000000000000000000120000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000014E6831e169d77a861A0E71326AFA6d80bCC8Bc6aA0000000000000000000000000000000000000000000000000000000000000000000000000000000000000014E6831e169d77a861A0E71326AFA6d80bCC8Bc6aA000000000000000000000000000000000000000000000000000000000000000000000000000000000000001474d5b8eacfeb0dadaaf66403f40e304b3ef968b300000000000000000000000000000000000000000000000000000000000000000000000000000000000000046d756e6f0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000046d756e6f00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001474d5b8eacfeb0dadaaf66403f40e304b3ef968b3000000000000000000000000"
+    })
   })
 })
