+++ ts-sdk/ts-sdk.nix
@@ -15,7 +15,7 @@ _: {
           pname = packageJson.name;
           inherit (packageJson) version;
           src = ./.;
-          npmDepsHash = "sha256-yc3gtJ7PARYzDWmALcbRwwxtrSEBqUztq6tGKM5bhvc=";
+          npmDepsHash = "sha256-w0LMXYz+yBJtBsN85bGvo4hkM0QQ3xYpDp2suUv4Xz0=";
           doCheck = true;
           checkPhase = ''
             npm run test
